+++ activesupport/lib/active_support/core_ext/range/each.rb
@@ -11,12 +11,44 @@ module ActiveSupport
 
     def step(n = 1, &block)
       ensure_iteration_allowed
-      super
+
+      case n
+      when ActiveSupport::Duration
+        duration_step(n, &block)
+      else
+        super
+      end
     end
 
     private
       def ensure_iteration_allowed
+        raise TypeError, "can't iterate beginless range" unless self.begin
         raise TypeError, "can't iterate from #{first.class}" if first.is_a?(TimeWithZone)
+      end
+
+      def duration_step(step, &block)
+        if block_given?
+          iterate(step, &block)
+
+          self
+        else
+          Enumerator.new do |yielder|
+            iterate(step) do |value|
+              yielder << value
+            end
+          end
+        end
+      end
+
+      def iterate(step, &block)
+        value, i = self.begin, 0
+
+        while cover?(value)
+          yield value
+
+          i += 1
+          value = self.begin + (step * i)
+        end
       end
   end
 end